+++ lib/config/configuration.go
@@ -2096,6 +2096,7 @@ func applyWindowsDesktopConfig(fc *FileConfig, cfg *servicecfg.Config) error {
 				BaseDN:          dc.BaseDN,
 				Filters:         dc.Filters,
 				LabelAttributes: dc.LabelAttributes,
+				RDPPort:         cmp.Or(dc.RDPPort, int(defaults.RDPListenPort)),
 			},
 		)
 	}